requirements.txt
# Testing
playwright==1.48.0
pytest==8.3.3
pytest-asyncio==0.24.0

# Telegram Bot
python-telegram-bot==21.4

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Logging
loguru==0.7.2

# Testing
playwright==1.48.0
pytest==8.3.3
pytest-asyncio==0.24.0

# Telegram Bot
python-telegram-bot==21.4

# Utilities
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.2